requirements.txt
streamlit==1.37.0
plotly==5.23.0
pandas
pillow
streamlit-plotly-events
